requirements.txt
# ── Core LLM & Agent Framework ──────────────────────────────────────────────
langchain>=0.3.0
langchain-nvidia-ai-endpoints>=0.3.0
langchain-community>=0.3.0
langgraph>=0.2.0
langchain-chroma>=0.2.0

# ── Vector Store & Embeddings ────────────────────────────────────────────────
chromadb>=0.5.0

# ── API & Web ────────────────────────────────────────────────────────────────
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.0
sse-starlette>=1.6.0
slowapi>=0.1.9
websockets>=12.0

# ── Frontend ─────────────────────────────────────────────────────────────────
streamlit>=1.38.0
streamlit-mermaid>=0.2.0

# ── Database ─────────────────────────────────────────────────────────────────
aiosqlite>=0.20.0

# ── Resilience & Retry ───────────────────────────────────────────────────────
tenacity>=8.2.0

# ── Scheduling ───────────────────────────────────────────────────────────────
apscheduler>=3.10.0

# ── Observability ────────────────────────────────────────────────────────────
python-json-logger>=2.0.0
langfuse>=2.0.0

# ── Security ─────────────────────────────────────────────────────────────────
cryptography>=42.0.0
python-jose[cryptography]>=3.3.0

# ── Evaluation ───────────────────────────────────────────────────────────────
ragas>=0.2.0
datasets>=3.0.0

# ── Testing ──────────────────────────────────────────────────────────────────
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.27.0

# ── Utilities ────────────────────────────────────────────────────────────────
python-dotenv>=1.0.0
python-multipart>=0.0.9
